# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  models.UpdateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
  models.VerifyEmailRequest:
    properties:
      token:
        type: string
    type: object
  response.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: suporte@example.com
    name: Suporte
  description: API de prova de conceito com autenticação JWT, MySQL e envio de email.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Go REST API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentica com email e senha e retorna access + refresh tokens.
      parameters:
      - description: Credenciais
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login realizado com sucesso
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Credenciais inválidas
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Login
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Gera novos access + refresh tokens a partir de um refresh token
        válido.
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tokens renovados com sucesso
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Refresh token inválido ou expirado
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Renovar tokens
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Cria um novo usuário e retorna os dados + tokens JWT. Um email
        de verificação é enviado.
      parameters:
      - description: Dados do registro
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Usuário criado com sucesso
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Erro de validação
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Email já cadastrado
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Registrar novo usuário
      tags:
      - auth
  /auth/verify-email:
    get:
      consumes:
      - application/json
      description: Marca o email do usuário como verificado usando o token enviado
        por email.
      parameters:
      - description: Token de verificação
        in: query
        name: token
        type: string
      - description: Token no body (alternativa)
        in: body
        name: request
        schema:
          $ref: '#/definitions/models.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Email verificado com sucesso
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Verificar email
      tags:
      - auth
    post:
      consumes:
      - application/json
      description: Marca o email do usuário como verificado usando o token enviado
        por email.
      parameters:
      - description: Token de verificação
        in: query
        name: token
        type: string
      - description: Token no body (alternativa)
        in: body
        name: request
        schema:
          $ref: '#/definitions/models.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Email verificado com sucesso
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/response.APIResponse'
      summary: Verificar email
      tags:
      - auth
  /users:
    get:
      description: Retorna a lista de todos os usuários cadastrados (requer autenticação).
      produces:
      - application/json
      responses:
        "200":
          description: Lista de usuários
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Não autorizado
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Listar usuários
      tags:
      - users
  /users/{id}:
    delete:
      description: Remove um usuário do sistema (requer role admin).
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Usuário deletado com sucesso
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Não autorizado
          schema:
            $ref: '#/definitions/response.APIResponse'
        "403":
          description: Acesso negado
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Usuário não encontrado
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Deletar usuário
      tags:
      - users
    get:
      description: Retorna os dados de um usuário específico (requer autenticação).
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Dados do usuário
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Não autorizado
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Usuário não encontrado
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Buscar usuário por ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Atualiza nome e/ou email de um usuário (requer autenticação).
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      - description: Dados para atualizar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Usuário atualizado com sucesso
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: ID ou corpo inválido
          schema:
            $ref: '#/definitions/response.APIResponse'
        "401":
          description: Não autorizado
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Usuário não encontrado
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Atualizar usuário
      tags:
      - users
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Insira o token JWT no formato: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"